refactor(providers): consolidate duplicated extraction helpers into openai_compat
vllm, llamacpp, and lmstudio each carried their own copy of
_extract_thinking_content / _extract_tool_calls_from_content (identical logic,
differing only in docstrings). Move canonical extract_thinking_content and
extract_tool_calls_from_content into victor.providers.openai_compat and have
the three providers alias them (preserving the private _extract_* names used
internally and by tests). mlx keeps its intentionally simpler tool-call parser.

Behavioral equivalence verified against the original implementations. The
duplicated TestExtractThinkingContent/TestExtractToolCallsFromContent test
classes are consolidated into test_openai_compat.py (one copy).
