+++ victor/providers/llamacpp_provider.py
@@ -61,6 +61,8 @@ from victor.providers.base import (
 )
 from victor.providers.logging import ProviderLogger
 from victor.providers.openai_compat import (
+    extract_thinking_content as _extract_thinking_content,
+    extract_tool_calls_from_content as _extract_tool_calls_from_content,
     convert_messages_to_openai_format,
     convert_tools_to_openai_format,
     parse_openai_tool_calls,
@@ -97,121 +99,6 @@ def _model_supports_tools(model: str) -> bool:
     """
     model_lower = model.lower()
     return any(pattern in model_lower for pattern in TOOL_CAPABLE_PATTERNS)
-
-
-def _extract_thinking_content(response: str) -> Tuple[str, str]:
-    """Extract thinking content from response.
-
-    Args:
-        response: Raw response text
-
-    Returns:
-        Tuple of (thinking_content, main_content)
-    """
-    think_pattern = r"<think>(.*?)</think>"
-    matches = re.findall(think_pattern, response, re.DOTALL | re.IGNORECASE)
-    thinking = "\n".join(matches) if matches else ""
-    content = re.sub(think_pattern, "", response, flags=re.DOTALL | re.IGNORECASE).strip()
-    return (thinking, content)
-
-
-def _extract_tool_calls_from_content(content: str) -> Tuple[List[Dict[str, Any]], str]:
-    """Extract tool calls from content when server doesn't parse them.
-
-    Handles cases where model outputs tool calls as JSON text.
-    Common formats:
-    - ```json\n{...}\n```
-    - <TOOL_OUTPUT>{...}</TOOL_OUTPUT>
-    - {"name": "...", "arguments": {...}}
-
-    Note: Must distinguish from task plan JSON which also has "name" but different structure.
-    Tool calls have "arguments" which are dict objects with tool-specific structure.
-
-    Args:
-        content: Response content that may contain tool calls
-
-    Returns:
-        Tuple of (parsed_tool_calls, remaining_content)
-    """
-    tool_calls = []
-    remaining = content
-
-    # Planning keywords to avoid false positives
-    planning_keywords = {"complexity", "steps", "desc", "duration"}
-
-    # Pattern 1: JSON code block with tool call
-    json_block_pattern = r"```json\s*\n?\s*(\{[^`]*\"name\"\s*:\s*\"[^\"]+\"[^`]*\})\s*\n?```"
-    matches = re.findall(json_block_pattern, content, re.DOTALL)
-    for match in matches:
-        try:
-            data = json.loads(match)
-            if "name" in data:
-                arguments = data.get("arguments", {})
-                # Only treat as tool call if arguments is a dict and doesn't look like planning JSON
-                if isinstance(arguments, dict) and not any(
-                    key in arguments for key in planning_keywords
-                ):
-                    tool_calls.append(
-                        {
-                            "id": f"fallback_{len(tool_calls)}",
-                            "name": data.get("name", ""),
-                            "arguments": arguments,
-                        }
-                    )
-                    remaining = remaining.replace(f"```json\n{match}\n```", "").strip()
-                    remaining = remaining.replace(f"```json{match}```", "").strip()
-        except json.JSONDecodeError:
-            pass
-
-    # Pattern 2: <TOOL_OUTPUT> tags
-    tool_output_pattern = r"<TOOL_OUTPUT>\s*(\{.*?\})\s*</TOOL_OUTPUT>"
-    matches = re.findall(tool_output_pattern, content, re.DOTALL)
-    for match in matches:
-        try:
-            data = json.loads(match)
-            if "name" in data:
-                arguments = data.get("arguments", {})
-                # Only treat as tool call if arguments is a dict and doesn't look like planning JSON
-                if isinstance(arguments, dict) and not any(
-                    key in arguments for key in planning_keywords
-                ):
-                    tool_calls.append(
-                        {
-                            "id": f"fallback_{len(tool_calls)}",
-                            "name": data.get("name", ""),
-                            "arguments": arguments,
-                        }
-                    )
-                    remaining = re.sub(
-                        r"<TOOL_OUTPUT>\s*" + re.escape(match) + r"\s*</TOOL_OUTPUT>",
-                        "",
-                        remaining,
-                    )
-        except json.JSONDecodeError:
-            pass
-
-    # Pattern 3: Inline JSON (for simple cases)
-    if not tool_calls and content.strip().startswith("{") and "name" in content:
-        try:
-            data = json.loads(content.strip())
-            if "name" in data:
-                arguments = data.get("arguments", {})
-                # Only treat as tool call if arguments is a dict and doesn't look like planning JSON
-                if isinstance(arguments, dict) and not any(
-                    key in arguments for key in planning_keywords
-                ):
-                    tool_calls.append(
-                        {
-                            "id": "fallback_0",
-                            "name": data.get("name", ""),
-                            "arguments": arguments,
-                        }
-                    )
-                    remaining = ""
-        except json.JSONDecodeError:
-            pass
-
-    return tool_calls, remaining.strip()
 
 
 class LlamaCppProvider(BaseProvider):
